requirements.txt
pandas
pillow
matplotlib
tensorflow